- optmize EFRepositoryBase.cs - Optmize GytDataService take code DataService easy
