+++ src/Zer.GytDataService/GytRepository.cs
@@ -1,0 +1,13 @@
+using Zer.Framework.Entities;
+using Zer.Framework.EntityFramework;
+
+namespace Zer.GytDataService
+{
+    public class GytRepository<TEntity> : EfRepositoryBase<TEntity>
+        where TEntity : class, IEntity<int>
+    {
+        protected GytRepository() : base(new GytDbContext())
+        {
+        }
+    }
+}
