+++ src/Zer.GytDataService/Zer.GytDataService.csproj
@@ -48,6 +48,7 @@
   <ItemGroup>
     <Compile Include="DataServiceModule.cs" />
     <Compile Include="GytDbContext.cs" />
+    <Compile Include="GytRepository.cs" />
     <Compile Include="ICompanyInfoDataService.cs" />
     <Compile Include="Impl\CompanyInfoDataService.cs" />
     <Compile Include="Migrations\201708031417529_Init_Databases.cs" />
